Rename the snap to minio-client (#2243)
In linux, mc is midnight commander.
In the traditional linux distro world, name conflicts are forbidden. Because we are trying to make this decentralized and self-serving, we need less strict rules. For this, we added the concept of aliases. Then the user can decide what application owns the mc name.

You could make a claim for the mc name, and keep it. But I think it would be less confusing for existing linux users to rename the snap and use mc as an alias.

I would love your feedback about this feature. Please let me know what do you think. Also, this needs somebody from your org to register the new name in build.snapcraft.io to get the continuous delivery working.